Rename to app.js as per standard practice.
