+++ run_server.sh
@@ -2,5 +2,5 @@
 
 forever stopall
 sudo /etc/init.d/varnish restart
-nohup forever server/index.js > output.log 2> error.log < /dev/null &
+nohup forever server/app.js > output.log 2> error.log < /dev/null &
